requirements.txt

python-dotenv==1.0.1
openai==1.12.0
pandas==2.2.0
pyarrow==15.0.0
numpy==1.26.4
torch==2.2.0
matplotlib==3.8.3
transformerstransformers==4.37.2
tqdm==4.66.2
scikit-learn==1.4.1.post1